backend/app/requirements.txt
streamlit
python-dotenv
langchain-core
langchain-community
huggingface-hub
flask
flask-sqlalchemy
bcrypt
pyjwt
langchain

# Flask==2.3.3
# requests==2.31.0
# gunicorn==21.2.0  # Optional for deploying the app
